requirements.txt
selenium
undetected-chromedriver
webdriver-manager
python-dotenv
setuptools 
requests